kvm tools: Add vhost-blk support
vhost-blk is an in-kernel virito-blk device accelerator. vhost-blk is
similar with vhost-net. It handles virito-blk's request and completion
in host kernel side.

How to use:
-----------------------------
Load the vhost-blk.ko module in host side and specify the vhost flag.
   $ lkvm run -d disk.img,vhost

Performance evaluation:
-----------------------------
The comparison is between kvm tool with usersapce implementation and kvm
tool with vhost-blk.

    Fio with libaio ioengine on Fusion IO device
    IOPS       Before   After   Improvement
    seq-read   107      121     +13.0%
    seq-write  130      179     +37.6%
    rnd-read   102      122     +19.6%
    rnd-write  125      159     +27.0%

Signed-off-by: Asias He <asias.hejun@gmail.com>